Refactor the create module
